requirements.txt
functools
git+ssh://git@github.com/wearefair/protobuf-to-dict
Flask
grpcio
grpcio-tools
